refactor: extract thread summary maintenance
